Create README.md
As an aspiring data analyst I decided to analyzed dataset on customer shopping data base the data was retrieved from kaggle.com. during this process I noticed age need to be grouped conditional column was applied on power query by grouping the "Age category", each if [age] <= 30 then "YOUTH" else if [age] <= 50 then "ADULT" else "OLD") by this I was able to learn some M language like (IF) and Switch. in terms of shopping clothing and shoe which was purchased higher among the FEMALE AND MALE.
Your suggestion and correction will be appreciated
below is the link to interact with the data.